Fix 32-bit pointer truncation casts, harden user shell stack, start net poll task
main.c/gpu.c: casting a uint32_t physical/MMIO address straight to a
pointer type triggers -Wint-to-pointer-cast noise (and is technically
wrong) once the codebase started mixing width-sensitive types elsewhere;
route through (uintptr_t) first.

main.c: allocate the ring-3 shell's stack via pmm_alloc_block() instead
of kmalloc(4096) — kmalloc'd memory lives in the kernel heap's virtual
range, so mapping it user+RW (flag 0x07) for the stack would expose
that range to ring-3; a dedicated physical page avoids that. Also start
net_poll_task() via task_create() before jumping to the ring-3 shell,
so networking keeps making progress as a background task once the
scheduler is running.

+++ kernel/gpu.c
@@ -271,7 +271,7 @@ void gpu_init() {
                 klog("GPU: NVIDIA BAR0 MMIO successfully mapped.");
                 
                 // Safe read of NV_PMC_BOOT_0
-                volatile uint32_t *pmc_boot_ptr = (volatile uint32_t *)(NVIDIA_BAR0_VMEM + NV_PMC_BOOT_0);
+                volatile uint32_t *pmc_boot_ptr = (volatile uint32_t *)(uintptr_t)(NVIDIA_BAR0_VMEM + NV_PMC_BOOT_0);
                 active_gpu.pmc_boot_0 = *pmc_boot_ptr;
                 
                 // Decode architecture family
